+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Changed F64.GetHashCode() to use System.Int64.GetHashCode().
 - Fixed64.Mod() checks for div-by-zero. Simplify the code by using the modulo operator (%).
+- Fixed all Pow*(X, 0) to always return precisely 1.0, including Pow*(0, 0) which previously returned 0.
 
 ## 0.1 (2019-06-22)